% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pass a Result to the Next Step}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The result to pass forward.}

\item{rhs}{The next function to run.}
}
\value{
The result from the function on the right.
}
\description{
\verb{\%>\%} passes the result on its left to the function on its right. See
\code{magrittr::\link[magrittr:pipe]{\%>\%}} for more examples.
}
\keyword{internal}
