% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RSYC_params}
\alias{RSYC_params}
\title{Parameters for the RSYC models}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1892 rows and 9 columns.
}
\usage{
RSYC_params
}
\description{
A data frame with 1892 rows and 9 columns:
\describe{
\item{TileID}{Tile identified}
\item{SpeciesCode, SpeciesName}{Species}
\item{b1_fixed, b1_random, b2, b3, b4}{model parameters}
...
}
}
\keyword{datasets}
