% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-catalog.R
\name{rsyc_species}
\alias{rsyc_species}
\title{List Species Available in RSYC Models}
\usage{
rsyc_species(response = NULL, strata_level = NULL, strata_id = NULL)
}
\arguments{
\item{response}{One or both forest measures: \code{"agb"} for aboveground biomass
and \code{"volume"} for total volume. The default includes both.}

\item{strata_level}{One or more types of area: \code{"tile"}, \code{"ecozone"},
\code{"ecoprovince"}, \code{"ecoregion"}, or \code{"ecodistrict"}. The default includes
all types.}

\item{strata_id}{One or more tile IDs, ecological area names, or full
sequences of ecological area names. The default includes all areas.}
}
\value{
A table giving each species code, common name, and broad model group.
}
\description{
List Species Available in RSYC Models
}
\examples{
rsyc_species(response = "agb", strata_level = "ecoregion")

}
