% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-catalog.R
\name{rsyc_strata}
\alias{rsyc_strata}
\title{List Areas Available in RSYC Models}
\usage{
rsyc_strata(response = NULL, species = NULL, strata_level = NULL)
}
\arguments{
\item{response}{One or both forest measures: \code{"agb"} for aboveground biomass
and \code{"volume"} for total volume. The default includes both.}

\item{species}{One or more species codes or broad groups. Uppercase and
lowercase letters are treated the same. The default includes all species.}

\item{strata_level}{One or more types of area: \code{"tile"}, \code{"ecozone"},
\code{"ecoprovince"}, \code{"ecoregion"}, or \code{"ecodistrict"}. The default includes
all types.}
}
\value{
A table with the short area name (\code{strata_name}) and its full
location within Canada's ecological classification (\code{strata_id}).
}
\description{
List Areas Available in RSYC Models
}
\examples{
rsyc_strata(
  response = "volume", species = "PICE.MAR", strata_level = "ecozone"
)

}
