% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{RSYC_tiles}
\alias{RSYC_tiles}
\title{RSYC Tile Grid (150 x 150 km)}
\description{
A GeoPackage containing the tile polygons used by the RSYC models. The path
can be obtained with \code{system.file("extdata", "RSYC_tiles.gpkg", package = "RSYC")} and read with \code{sf::st_read()}.
}
