% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\name{species_codes}
\alias{species_codes}
\title{List Available Species Codes and Names}
\usage{
species_codes(response = NULL, strata_level = NULL)
}
\arguments{
\item{response}{Forest measure to include: \code{"agb"} for aboveground biomass
or \code{"volume"} for total volume. By default, both are included.}

\item{strata_level}{Type of area to include. By default, all types are
included.}
}
\value{
A table with the species code and common name.
}
\description{
List the species codes and common names represented in the published RSYC
curves. Broad groups are listed first, followed by individual species.
}
\examples{
species_codes()
species_codes(response = "volume", strata_level = "ecoregion")

}
