% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_in_tile.R
\name{species_in_tile}
\alias{species_in_tile}
\title{List Species Available in a Tile}
\usage{
species_in_tile(tile_id, response = "agb")
}
\arguments{
\item{tile_id}{One tile ID.}

\item{response}{Forest measure: \code{"agb"} for aboveground biomass or
\code{"volume"} for total volume.}
}
\value{
The species codes and broad groups that have a published curve in
the chosen tile.
}
\description{
List Species Available in a Tile
}
\examples{
species_in_tile("H14")
species_in_tile("H14", response = "volume")

}
