% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RSYC_tiles}
\alias{RSYC_tiles}
\title{RSYC Tile Grid (150 × 150 km)}
\format{
An \code{sf} object with one feature per tile and the following attribute:
\describe{
\item{TileID}{Unique identifier of the 150 × 150 km tile (character).}
}
}
\usage{
RSYC_tiles
}
\description{
A spatial layer defining the 150 × 150 km tiles used in the development of
the Remote Sensing Yield Curves (RSYC). Each polygon represents one tile
for which yield curves were developed. The layer can be used to identify the
models to use in given location.
}
\examples{
# Load the layer
library(sf)
data(RSYC_tiles)

# Plot
plot(st_geometry(RSYC_tiles))

}
\keyword{datasets}
