% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\name{species_codes}
\alias{species_codes}
\title{List Available Species Codes and Names}
\usage{
species_codes()
}
\value{
A tibble with two columns: \code{SpeciesCode} and \code{SpeciesName}.
}
\description{
This function returns a data frame of species codes and their corresponding species names
used in the RSYC model set. It arranges species codes alphabetically, but promotes the
meta-categories \code{"Generic"}, \code{"Broadleaf"}, and \code{"Coniferous"} to the top of the list.
}
\examples{
species_codes()

}
