% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_in_tile.R
\name{species_in_tile}
\alias{species_in_tile}
\title{List Species Available in a Tile}
\usage{
species_in_tile(tile_id, response = "agb")
}
\arguments{
\item{tile_id}{A single tile ID.}

\item{response}{Response variable: \code{"agb"} or \code{"volume"}.}
}
\value{
A character vector of canonical species codes and model groups.
}
\description{
List Species Available in a Tile
}
\examples{
species_in_tile("H14")
species_in_tile("H14", response = "volume")

}
