% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_in_tile.R
\name{species_in_tile}
\alias{species_in_tile}
\title{List Available Species for a Given Tile}
\usage{
species_in_tile(tile_id)
}
\arguments{
\item{tile_id}{A character string specifying the tile ID.}
}
\value{
A character vector of species codes available for the given tile.
}
\description{
This function returns the species codes for which RSYC models are available within a specified tile.
}
\examples{
species_in_tile("B22")

}
