% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_for_species.R
\name{tiles_for_species}
\alias{tiles_for_species}
\title{List Tiles Available for a Species}
\usage{
tiles_for_species(species, response = "agb")
}
\arguments{
\item{species}{A single species code or model group. Matching is
case-insensitive.}

\item{response}{Response variable: \code{"agb"} or \code{"volume"}.}
}
\value{
A character vector of tile IDs.
}
\description{
List Tiles Available for a Species
}
\examples{
tiles_for_species("PINU.CON")
tiles_for_species("Coniferous", response = "volume")

}
