% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_for_species.R
\name{tiles_for_species}
\alias{tiles_for_species}
\title{List Available Tiles for a Given Species}
\usage{
tiles_for_species(species)
}
\arguments{
\item{species}{A character string specifying the species code (e.g., "PICE.MAR" for black spruce).}
}
\value{
A character vector of tile IDs where the specified species is available.
}
\description{
This function returns the tile IDs where RSYC models are available for a specified species.
}
\examples{
tiles_for_species("PINU.CON")

}
